Adding sha256 digest for each of the exisiting charts. (#214)
We are adding sha256 digest as an additional chart field. This matches
what helm index command does and it improves our capabilities to do
integrity checking. In Addition this will fix a gui dependency on digest
to use as a unique ID for the charts. A different PR will be sent to
compute this digest as part of the ci flow.