Adding the PostgreSQL Service Endpoint Definition (SED) chart.
This chart can be used to point to a PostgreSQL database. The result
when installing this chart is a secret the contains the binding
data compliant with the Service Binding spec 1.0.0. The secret can
be used to project the binding data to workloads referenced by
ServiceBinding CRs. This will be useful when a developer is testing
SBO with services deliverd via helm charts which do not contain
CRDs representing provisoned service.